# app/schemas/generated.py
# Code generated by datamodel-codegen from openapi/spec.yaml
# DO NOT EDIT MANUALLY — regenerate with:
#   uv run datamodel-codegen \
#       --input openapi/spec.yaml --input-file-type openapi \
#       --output app/schemas/generated.py \
#       --target-python-version 3.12 --use-annotated \
#       --field-constraints --use-standard-collections --use-union-operator
#
# After regeneration: ensure Product and OrderItem retain
# model_config = ConfigDict(from_attributes=True)
from __future__ import annotations

from datetime import datetime
from enum import Enum
from typing import Annotated, Literal

from pydantic import BaseModel, ConfigDict, Field


class ProductCategory(str, Enum):
    electronics = "electronics"
    clothing = "clothing"
    food = "food"
    books = "books"
    sports = "sports"
    other = "other"


class OrderStatus(str, Enum):
    pending = "pending"
    confirmed = "confirmed"
    processing = "processing"
    shipped = "shipped"
    delivered = "delivered"
    cancelled = "cancelled"


class HealthResponse(BaseModel):
    status: Literal["ok", "degraded"]
    version: str
    uptime_seconds: float


class ProductBase(BaseModel):
    name: Annotated[str, Field(min_length=1, max_length=255)]
    description: str | None = None
    price: Annotated[float, Field(ge=0.01)]
    stock_quantity: Annotated[int, Field(ge=0)] = 0
    category: ProductCategory
    sku: str | None = None


class ProductCreate(ProductBase):
    pass


class ProductUpdate(BaseModel):
    name: Annotated[str, Field(min_length=1, max_length=255)] | None = None
    description: str | None = None
    price: Annotated[float, Field(ge=0.01)] | None = None
    stock_quantity: Annotated[int, Field(ge=0)] | None = None
    category: ProductCategory | None = None
    sku: str | None = None


class Product(ProductBase):
    model_config = ConfigDict(from_attributes=True)

    id: int
    created_at: datetime
    updated_at: datetime


class ProductListResponse(BaseModel):
    items: list[Product]
    total: int
    limit: int
    offset: int


class OrderItemCreate(BaseModel):
    product_id: int
    quantity: Annotated[int, Field(ge=1)]


class OrderItem(BaseModel):
    model_config = ConfigDict(from_attributes=True)

    id: int
    product_id: int
    product_name: str | None = None
    quantity: int
    unit_price: float


class OrderCreate(BaseModel):
    customer_id: Annotated[str, Field(min_length=1)]
    items: Annotated[list[OrderItemCreate], Field(min_length=1)]
    notes: str | None = None


class OrderStatusUpdate(BaseModel):
    status: OrderStatus


class Order(BaseModel):
    id: int
    customer_id: str
    status: OrderStatus
    total_amount: float
    items: list[OrderItem]
    notes: str | None = None
    created_at: datetime
    updated_at: datetime


class OrderListResponse(BaseModel):
    items: list[Order]
    total: int
    limit: int
    offset: int


class ErrorResponse(BaseModel):
    error: str
    message: str
    details: dict | None = None