# openapi/spec.yaml
openapi: 3.0.3

info:
  title: Store Service API
  description: |
    Product catalog and order management service.
    Provides CRUD for products and a state-machine-driven order lifecycle.
    Designed for observability: all error paths carry structured `error` codes
    suitable for ELK ingestion.
  version: 1.0.0

servers:
  - url: http://localhost:8000
    description: Development

tags:
  - name: health
    description: Liveness / readiness probes
  - name: products
    description: Product catalog
  - name: orders
    description: Order lifecycle management

paths:
  /health:
    get:
      tags: [health]
      summary: Health check
      operationId: health_check
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"

  /products:
    get:
      tags: [products]
      summary: List products
      operationId: list_products
      parameters:
        - name: category
          in: query
          schema:
            $ref: "#/components/schemas/ProductCategory"
        - name: min_price
          in: query
          schema:
            type: number
            minimum: 0
        - name: max_price
          in: query
          schema:
            type: number
            minimum: 0
        - name: in_stock
          in: query
          schema:
            type: boolean
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        "200":
          description: Paginated product list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"

    post:
      tags: [products]
      summary: Create a product
      operationId: create_product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductCreate"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"

  /products/{product_id}:
    parameters:
      - name: product_id
        in: path
        required: true
        schema:
          type: integer

    get:
      tags: [products]
      summary: Get product by ID
      operationId: get_product
      responses:
        "200":
          description: Product details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      tags: [products]
      summary: Update product
      operationId: update_product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"

    delete:
      tags: [products]
      summary: Delete product
      operationId: delete_product
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"

  /orders:
    get:
      tags: [orders]
      summary: List orders
      operationId: list_orders
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/OrderStatus"
        - name: customer_id
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        "200":
          description: Paginated order list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListResponse"

    post:
      tags: [orders]
      summary: Create an order
      operationId: create_order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderCreate"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"

  /orders/{order_id}:
    parameters:
      - name: order_id
        in: path
        required: true
        schema:
          type: integer

    get:
      tags: [orders]
      summary: Get order by ID
      operationId: get_order
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      tags: [orders]
      summary: Cancel order
      operationId: cancel_order
      responses:
        "200":
          description: Cancelled order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"

  /orders/{order_id}/status:
    parameters:
      - name: order_id
        in: path
        required: true
        schema:
          type: integer

    patch:
      tags: [orders]
      summary: Advance order status
      operationId: update_order_status
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Order with updated status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"

components:
  schemas:
    HealthResponse:
      type: object
      required: [status, version, uptime_seconds]
      properties:
        status:
          type: string
          enum: [ok, degraded]
        version:
          type: string
        uptime_seconds:
          type: number

    ProductCategory:
      type: string
      enum:
        - electronics
        - clothing
        - food
        - books
        - sports
        - other

    ProductBase:
      type: object
      required: [name, price, category]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          nullable: true
        price:
          type: number
          format: float
          minimum: 0.01
        stock_quantity:
          type: integer
          minimum: 0
          default: 0
        category:
          $ref: "#/components/schemas/ProductCategory"
        sku:
          type: string
          nullable: true
          maxLength: 100

    ProductCreate:
      allOf:
        - $ref: "#/components/schemas/ProductBase"

    ProductUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          nullable: true
        price:
          type: number
          format: float
          minimum: 0.01
        stock_quantity:
          type: integer
          minimum: 0
        category:
          $ref: "#/components/schemas/ProductCategory"
        sku:
          type: string
          nullable: true
          maxLength: 100

    Product:
      allOf:
        - $ref: "#/components/schemas/ProductBase"
        - type: object
          required: [id, created_at, updated_at]
          properties:
            id:
              type: integer
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time

    ProductListResponse:
      type: object
      required: [items, total, limit, offset]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer

    OrderStatus:
      type: string
      enum:
        - pending
        - confirmed
        - processing
        - shipped
        - delivered
        - cancelled

    OrderItemCreate:
      type: object
      required: [product_id, quantity]
      properties:
        product_id:
          type: integer
        quantity:
          type: integer
          minimum: 1

    OrderItem:
      type: object
      required: [id, product_id, quantity, unit_price]
      properties:
        id:
          type: integer
        product_id:
          type: integer
        product_name:
          type: string
          nullable: true
        quantity:
          type: integer
        unit_price:
          type: number
          format: float

    OrderCreate:
      type: object
      required: [customer_id, items]
      properties:
        customer_id:
          type: string
          minLength: 1
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemCreate"
          minItems: 1
        notes:
          type: string
          nullable: true

    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          $ref: "#/components/schemas/OrderStatus"

    Order:
      type: object
      required: [id, customer_id, status, total_amount, items, created_at, updated_at]
      properties:
        id:
          type: integer
        customer_id:
          type: string
        status:
          $ref: "#/components/schemas/OrderStatus"
        total_amount:
          type: number
          format: float
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        notes:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    OrderListResponse:
      type: object
      required: [items, total, limit, offset]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer

    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
          nullable: true

  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

    Conflict:
      description: Business rule violation or resource conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

    ValidationError:
      description: Request payload validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"